# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: MoMo APIs
  description: HTTP Service for accessing MTN's MoMo API.
  termsOfService: https://github.com/wondenge/momo-go/blob/master/LICENCE
  contact:
    name: William Ondenge
    email: ondengew@gmail.com
    url: https://www.ondenge.me
  license:
    name: Apache License
    url: https://github.com/wondenge/momo-go/blob/master/LICENCE
  version: "1.0"
host: sandbox.momodeveloper.mtn.com
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /collection/token:
    post:
      tags:
      - collection
      summary: NewToken collection
      description: Creates an Access Token.
      operationId: collection#NewToken
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CollectionNewTokenResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/CollectionNewTokenUnauthorizedResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CollectionNewTokenInternalErrorResponseBody'
      schemes:
      - https
  /collection/v1_0/account/balance:
    get:
      tags:
      - collection
      summary: GetBalance collection
      description: Get the balance of the account
      operationId: collection#GetBalance
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/CollectionGetBalanceResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/CollectionGetBalanceBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CollectionGetBalanceInternalErrorResponseBody'
      schemes:
      - https
  /collection/v1_0/accountholder/{accountHolderIdType}/{accountHolderId}/active:
    get:
      tags:
      - collection
      summary: RetrieveAccountHolder collection
      description: Checks if an account holder is registered and active in the system
      operationId: collection#RetrieveAccountHolder
      parameters:
      - name: accountHolderIdType
        in: path
        description: Specifies the type of the party ID
        required: true
        type: string
      - name: accountHolderId
        in: path
        description: The party number.
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/CollectionRetrieveAccountHolderBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/CollectionRetrieveAccountHolderInternalErrorResponseBody'
      schemes:
      - https
...
  /disbursement/token:
    post:
      tags:
      - disbursement
      summary: NewToken disbursement
      description: Creates an Access Token.
      operationId: disbursement#NewToken
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/DisbursementNewTokenResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/DisbursementNewTokenUnauthorizedResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/DisbursementNewTokenInternalErrorResponseBody'
      schemes:
      - https
  /disbursement/v1_0/account/balance:
    get:
      tags:
      - disbursement
      summary: GetBalance disbursement
      description: Get the balance of the account
      operationId: disbursement#GetBalance
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/DisbursementGetBalanceResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/DisbursementGetBalanceBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/DisbursementGetBalanceInternalErrorResponseBody'
      schemes:
      - https
  /disbursement/v1_0/accountholder/{accountHolderIdType}/{accountHolderId}/active:
    get:
      tags:
      - disbursement
      summary: RetrieveAccountHolder disbursement
      description: Checks if an account holder is registered and active in the system
      operationId: disbursement#RetrieveAccountHolder
      parameters:
      - name: accountHolderIdType
        in: path
        description: Specifies the type of the party ID
        required: true
        type: string
      - name: accountHolderId
        in: path
        description: The party number.
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/DisbursementRetrieveAccountHolderBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/DisbursementRetrieveAccountHolderInternalErrorResponseBody'
      schemes:
      - https
  /disbursement/v1_0/transfer:
    post:
      tags:
      - disbursement
      summary: Transfer disbursement
      description: transfer an amount from the owner’s account to a payee account.
      operationId: disbursement#Transfer
      parameters:
      - name: TransferRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/DisbursementTransferRequestBody'
      responses:
        "202":
          description: Accepted response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/DisbursementTransferBadRequestResponseBody'
        "409":
          description: Conflict response.
          schema:
            $ref: '#/definitions/DisbursementTransferConflictResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/DisbursementTransferInternalErrorResponseBody'
      schemes:
      - https
  /disbursement/v1_0/transfer/{referenceId}:
    get:
      tags:
      - disbursement
      summary: TransferStatus disbursement
      description: Get the status of a transfer.
      operationId: disbursement#TransferStatus
      produces:
      - application/json
      parameters:
      - name: referenceId
        in: path
        description: ' UUID of transaction to get result'
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/DisbursementTransferStatusResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/DisbursementTransferStatusBadRequestResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/DisbursementTransferStatusNotFoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/DisbursementTransferStatusInternalErrorResponseBody'
      schemes:
      - https
  /remittance/token:
    post:
      tags:
      - remittance
      summary: NewToken remittance
      description: Creates an Access Token.
      operationId: remittance#NewToken
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/RemittanceNewTokenResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/RemittanceNewTokenUnauthorizedResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/RemittanceNewTokenInternalErrorResponseBody'
      schemes:
      - https
  /remittance/v1_0/account/balance:
    get:
      tags:
      - remittance
      summary: GetBalance remittance
      description: Get the balance of the account
      operationId: remittance#GetBalance
      produces:
      - application/json
      parameters:
      - name: string
        in: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/RemittanceGetBalanceResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/RemittanceGetBalanceBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/RemittanceGetBalanceInternalErrorResponseBody'
      schemes:
      - https
  /remittance/v1_0/accountholder/{accountHolderIdType}/{accountHolderId}/active:
    get:
      tags:
      - remittance
      summary: RetrieveAccountHolder remittance
      description: Checks if an account holder is registered and active in the system
      operationId: remittance#RetrieveAccountHolder
      parameters:
      - name: accountHolderIdType
        in: path
        description: Specifies the type of the party ID
        required: true
        type: string
      - name: accountHolderId
        in: path
        description: The party number.
        required: true
        type: string
      responses:
        "200":
          description: Ok. True if account holder is registered and active, false if the account holder is not active or not found found
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/RemittanceRetrieveAccountHolderBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/RemittanceRetrieveAccountHolderInternalErrorResponseBody'
      schemes:
      - https
  /remittance/v1_0/transfer:
    post:
      tags:
      - remittance
      summary: Transfer remittance
      description: Request a payment from a consumer (Payer).
      operationId: remittance#Transfer
      parameters:
      - name: TransferRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/RemittanceTransferRequestBody'
      responses:
        "202":
          description: Accepted response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/RemittanceTransferBadRequestResponseBody'
        "409":
          description: Conflict response.
          schema:
            $ref: '#/definitions/RemittanceTransferConflictResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/RemittanceTransferInternalErrorResponseBody'
      schemes:
      - https
  /remittance/v1_0/transfer/{referenceId}:
    get:
      tags:
      - remittance
      summary: TransferStatus remittance
      description: Get the status of a request to pay.
      operationId: remittance#TransferStatus
      produces:
      - application/json
      parameters:
      - name: referenceId
        in: path
        description: ' UUID of transaction to get result'
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/RemittanceTransferStatusResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/RemittanceTransferStatusBadRequestResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/RemittanceTransferStatusNotFoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/RemittanceTransferStatusInternalErrorResponseBody'
      schemes:
      - https
...
definitions:
  CollectionGetBalanceBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Bad request, e.g. invalid data was sent in the request. (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CollectionGetBalanceInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Internal error. The returned response contains details. (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CollectionGetBalanceResponseBody:
    title: 'Mediatype identifier: balance; view=default'
    type: object
    properties:
      availableBalance:
        type: string
        description: The available balance of the account
        example: Vel in autem.
      currency:
        type: string
        description: ISO4217 Currency
        example: Unde aut dolor aut consequatur.
    description: GetBalanceResponseBody result type (default view)
    example:
      availableBalance: Error qui asperiores repudiandae blanditiis illum quos.
      currency: Provident ea placeat rem.
  CollectionNewTokenInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Error (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  CollectionNewTokenResponseBody:
    title: 'Mediatype identifier: tokenpost200applicationjsonresponse; view=default'
    type: object
    properties:
      access_token:
        type: string
        description: A JWT token which can be used to authorize against the other API end-points.
        example: Iusto beatae est ut perspiciatis ut harum.
      expires_in:
        type: integer
        description: The validity time in seconds of the token.
        example: 1892691476
        format: int32
      token_type:
        type: string
        description: The token type.
        example: Aspernatur ut et consequuntur placeat quam.
    description: NewTokenResponseBody result type (default view)
    example:
      access_token: Est dolores ipsa.
      expires_in: 1590797597
      token_type: Minima numquam quas omnis quis deserunt cupiditate.
  CollectionNewTokenUnauthorizedResponseBody:
    title: 'Mediatype identifier: tokenpost401applicationjsonresponse; view=default'
    type: object
    properties:
      error:
        type: string
        description: An error code.
        example: Veniam sit doloribus qui ea nobis libero.
    description: Unauthorized (default view)
    example:
      error: Qui est cupiditate autem dolorum nostrum ut.
...
  CollectionRetrieveAccountHolderBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Bad request, e.g. invalid data was sent in the request. (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  DisbursementNewTokenResponseBody:
    title: 'Mediatype identifier: tokenpost200applicationjsonresponse; view=default'
    type: object
    properties:
      access_token:
        type: string
        description: A JWT token which can be used to authorize against the other API end-points.
        example: Similique nam quas.
      expires_in:
        type: integer
        description: The validity time in seconds of the token.
        example: 1650204335
        format: int32
      token_type:
        type: string
        description: The token type.
        example: Est laborum in ratione.
    description: NewTokenResponseBody result type (default view)
    example:
      access_token: Odio itaque id qui laborum repudiandae et.
      expires_in: 1061546812
      token_type: Maxime sit tempore quasi cupiditate.
...